Move ZeroBasedIndexConverter to OmniSharp.Models
