+++ src/OmniSharp.Abstractions/Models/Request.cs
@@ -1,7 +1,6 @@
 ﻿using System.Collections.Generic;
 using System.IO;
 using Newtonsoft.Json;
-using OmniSharp.Json;
 
 namespace OmniSharp.Models
 {
