+++ src/OmniSharp.Abstractions/Models/ZeroBasedIndexConverter.cs
@@ -3,7 +3,7 @@ using System.Collections.Generic;
 using System.Linq;
 using Newtonsoft.Json;
 
-namespace OmniSharp.Json
+namespace OmniSharp.Models
 {
     public class ZeroBasedIndexConverter : JsonConverter
     {
