+++ src/OmniSharp.Abstractions/Models/v1/LinePositionSpanTextChange.cs
@@ -4,7 +4,6 @@ using System.Threading.Tasks;
 using Microsoft.CodeAnalysis;
 using Microsoft.CodeAnalysis.Text;
 using Newtonsoft.Json;
-using OmniSharp.Json;
 
 namespace OmniSharp.Models
 {
