+++ src/OmniSharp.Abstractions/Models/v1/Navigate/NavigateResponse.cs
@@ -1,6 +1,4 @@
 using Newtonsoft.Json;
-using OmniSharp.Json;
-
 namespace OmniSharp.Models.Navigate
 {
     public class NavigateResponse
