+++ src/OmniSharp.Abstractions/Models/v1/QuickFix.cs
@@ -1,6 +1,5 @@
 ﻿using System.Collections.Generic;
 using Newtonsoft.Json;
-using OmniSharp.Json;
 
 namespace OmniSharp.Models
 {
